move server and build to their own files
